+++ doc/app.js
@@ -1,3 +1,4 @@
+import '@babel/polyfill';
 import Vue from 'vue';
 import VueI18n from 'vue-i18n';
 import VueRouter from 'vue-router';
@@ -42,11 +43,10 @@ const i18n = new VueI18n({
   fallbackLocale: 'en',
   messages // set locale messages
 });
-HeyUI.i18n(i18n);
 
 heyuiConfig();
 Vue.use(VueRouter);
-Vue.use(HeyUI);
+Vue.use(HeyUI, { i18n });
 
 Vue.component('example', example);
 Vue.component('codes', codes);