+++ doc/components/demos/view/table5.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <p>
-      <hSwitch v-model="loading">loading</hSwitch>
+      <HSwitch v-model="loading">loading</HSwitch>
       <Button @click="updateSort">Age降序</Button>
     </p>
     <Table ref="table" :datas="datas" :loading="loading" :columns="columns" @sort="triggerSort">
@@ -17,6 +17,7 @@ export default {
       loading: false,
       columns: [
         { title: 'Index', prop: '$index', width: 100 },
+        { title: 'Serial', prop: '$serial', width: 100 },
         { title: 'ID', prop: 'id', width: 100, sort: true },
         { title: 'Name', prop: 'name', sort: true },
         { title: 'Age', render: item => `年龄：${item.age}`, sortProp: 'age', sort: true },