+++ src/components/table/tabletd.vue
@@ -1,0 +1,47 @@
+<template>
+  <td :class="cls"><template v-if="prop || render">{{show}}</template><slot :data="data" :index="index"></slot></td>
+</template>
+<script>
+import config from 'heyui/src/utils/config';
+import hutils from 'hey-utils';
+
+export default {
+  name: 'hTableTd',
+  props: {
+    index: Number,
+    prop: String,
+    dict: String,
+    data: [Object, Array],
+    align: String,
+    unit: String,
+    render: Function,
+    className: String
+  },
+  data() {
+    return {};
+  },
+  computed: {
+    cls() {
+      return {
+        [`text-${this.align}`]: !!this.align,
+        [this.className]: !!this.className
+      };
+    },
+    show() {
+      if (this.prop == '$index') return this.index;
+      if (this.prop == '$serial') return this.index + 1;
+      if (this.render) {
+        return this.render.call(null, this.data);
+      }
+      let value = hutils.getKeyValue(this.data, this.prop);
+      if (this.dict) {
+        return config.dictMapping(value, this.dict);
+      }
+      if (this.unit) {
+        return value === '' || value === null || value === undefined ? '' : `${value}${this.unit}`;
+      }
+      return value;
+    }
+  }
+};
+</script>