+++ src/directives/tooltip.js
@@ -1,5 +1,5 @@
-import Tooltip from '../plugins/tooltip';
-import utils from '../utils/utils';
+import Tooltip from 'heyui/src/plugins/tooltip';
+import utils from 'heyui/src/utils/utils';
 
 const getContent = function (el, vnode) {
   let param = {};